change icon address in Chart.yaml
the origin addresses of icons in Chart.yaml are from different website, and
several links are invlid at present. Now the related icons are uploaded to
tencent cdn service,this service is stable.  And the addresses are replaced
by adresses on cdn.

By the way ,move pular from incubator to community